LayoutTests/imported/w3c/web-platform-tests/css/css-text/white-space/seg-break-transformation-002.tentative-expected.txt


FAIL linebreak only assert_equals: expected 216 but got 223
FAIL spaces linebreak assert_equals: expected 216 but got 223
FAIL linebreak spaces assert_equals: expected 216 but got 223
FAIL spaces linebreak spaces assert_equals: expected 216 but got 223
FAIL multiple linebreaks assert_equals: expected 216 but got 223
FAIL multiple linebreaks + spaces assert_equals: expected 216 but got 223
ＦＵＬＬ ＷＩＤＴＨ
ＦＵＬＬ ＷＩＤＴＨ
ＦＵＬＬ ＷＩＤＴＨ
ＦＵＬＬ ＷＩＤＴＨ
ＦＵＬＬ ＷＩＤＴＨ
ＦＵＬＬ ＷＩＤＴＨ
ＦＵＬＬＷＩＤＴＨ
